+++ api/photos.js
@@ -1,5 +1,6 @@
 const { Router } = require('express')
 const { ValidationError } = require('sequelize')
+const { requireAuthentication, isAuthorized, unauthorizedResponse } = require('../lib/auth')
 
 const { Photo, PhotoClientFields } = require('../models/photo')
 
@@ -8,16 +9,20 @@ const router = Router()
 /*
  * Route to create a new photo.
  */
-router.post('/', async function (req, res, next) {
-  try {
-    const photo = await Photo.create(req.body, PhotoClientFields)
-    res.status(201).send({ id: photo.id })
-  } catch (e) {
-    if (e instanceof ValidationError) {
-      res.status(400).send({ error: e.message })
-    } else {
-      throw e
+router.post('/', requireAuthentication, async function (req, res, next) {
+  if (isAuthorized(req)) {
+    try {
+      const photo = await Photo.create(req.body, PhotoClientFields)
+      res.status(201).send({ id: photo.id })
+    } catch (e) {
+      if (e instanceof ValidationError) {
+        res.status(400).send({ error: e.message })
+      } else {
+        throw e
+      }
     }
+  } else {
+    unauthorizedResponse(res)
   }
 })
 
@@ -37,35 +42,43 @@ router.get('/:photoId', async function (req, res, next) {
 /*
  * Route to update a photo.
  */
-router.patch('/:photoId', async function (req, res, next) {
-  const photoId = req.params.photoId
-
-  /*
-   * Update photo without allowing client to update businessId or userId.
-   */
-  const result = await Photo.update(req.body, {
-    where: { id: photoId },
-    fields: PhotoClientFields.filter(
-      field => field !== 'businessId' && field !== 'userId'
-    )
-  })
-  if (result[0] > 0) {
-    res.status(204).send()
+router.patch('/:photoId', requireAuthentication, async function (req, res, next) {
+  if (isAuthorized(req)) {
+    const photoId = req.params.photoId
+  
+    /*
+     * Update photo without allowing client to update businessId or userId.
+     */
+    const result = await Photo.update(req.body, {
+      where: { id: photoId },
+      fields: PhotoClientFields.filter(
+        field => field !== 'businessId' && field !== 'userId'
+      )
+    })
+    if (result[0] > 0) {
+      res.status(204).send()
+    } else {
+      next()
+    }
   } else {
-    next()
+    unauthorizedResponse(res)
   }
 })
 
 /*
  * Route to delete a photo.
  */
-router.delete('/:photoId', async function (req, res, next) {
-  const photoId = req.params.photoId
-  const result = await Photo.destroy({ where: { id: photoId }})
-  if (result > 0) {
-    res.status(204).send()
+router.delete('/:photoId',requireAuthentication, async function (req, res, next) {
+  if (isAuthorized(req)) {
+    const photoId = req.params.photoId
+    const result = await Photo.destroy({ where: { id: photoId }})
+    if (result > 0) {
+      res.status(204).send()
+    } else {
+      next()
+    }
   } else {
-    next()
+    unauthorizedResponse(res)
   }
 })
