+++ api/reviews.js
@@ -1,5 +1,6 @@
 const { Router } = require('express')
 const { ValidationError } = require('sequelize')
+const { requireAuthentication, isAuthorized, unauthorizedResponse } = require('../lib/auth')
 
 const { Review, ReviewClientFields } = require('../models/review')
 
@@ -8,16 +9,20 @@ const router = Router()
 /*
  * Route to create a new review.
  */
-router.post('/', async function (req, res, next) {
-  try {
-    const review = await Review.create(req.body, ReviewClientFields)
-    res.status(201).send({ id: review.id })
-  } catch (e) {
-    if (e instanceof ValidationError) {
-      res.status(400).send({ error: e.message })
-    } else {
-      throw e
+router.post('/', requireAuthentication, async function (req, res, next) {
+  if (isAuthorized(req)) {
+    try {
+      const review = await Review.create(req.body, ReviewClientFields)
+      res.status(201).send({ id: review.id })
+    } catch (e) {
+      if (e instanceof ValidationError) {
+        res.status(400).send({ error: e.message })
+      } else {
+        throw e
+      }
     }
+  } else {
+    unauthorizedResponse(res)
   }
 })
 
@@ -37,35 +42,43 @@ router.get('/:reviewId', async function (req, res, next) {
 /*
  * Route to update a review.
  */
-router.patch('/:reviewId', async function (req, res, next) {
-  const reviewId = req.params.reviewId
-
-  /*
-   * Update review without allowing client to update businessId or userId.
-   */
-  const result = await Review.update(req.body, {
-    where: { id: reviewId },
-    fields: ReviewClientFields.filter(
-      field => field !== 'businessId' && field !== 'userId'
-    )
-  })
-  if (result[0] > 0) {
-    res.status(204).send()
+router.patch('/:reviewId', requireAuthentication, async function (req, res, next) {
+  if (isAuthorized(req)) {
+    const reviewId = req.params.reviewId
+  
+    /*
+     * Update review without allowing client to update businessId or userId.
+     */
+    const result = await Review.update(req.body, {
+      where: { id: reviewId },
+      fields: ReviewClientFields.filter(
+        field => field !== 'businessId' && field !== 'userId'
+      )
+    })
+    if (result[0] > 0) {
+      res.status(204).send()
+    } else {
+      next()
+    }
   } else {
-    next()
+    unauthorizedResponse(res)
   }
 })
 
 /*
  * Route to delete a review.
  */
-router.delete('/:reviewId', async function (req, res, next) {
-  const reviewId = req.params.reviewId
-  const result = await Review.destroy({ where: { id: reviewId }})
-  if (result > 0) {
-    res.status(204).send()
+router.delete('/:reviewId', requireAuthentication, async function (req, res, next) {
+  if (isAuthorized(req)) {
+    const reviewId = req.params.reviewId
+    const result = await Review.destroy({ where: { id: reviewId }})
+    if (result > 0) {
+      res.status(204).send()
+    } else {
+      next()
+    }
   } else {
-    next()
+    unauthorizedResponse(res)
   }
 })
